move components to src/components/
