+++ packages/components/test/components/Folder.test.tsx
@@ -3,7 +3,7 @@ import { render, waitFor } from '@testing-library/react'
 import { strict as assert } from 'assert'
 import React from 'react'
 import { describe, expect, it, vi } from 'vitest'
-import Folder from '../src/Folder.js'
+import Folder from '../../src/components/Folder.js'
 
 vi.mock('@hyparam/utils', async (importOriginal) => {
   const actual = await importOriginal()
