+++ packages/components/test/components/viewers/MarkdownView.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react'
 import { strict as assert } from 'assert'
 import React from 'react'
 import { describe, expect, it, vi } from 'vitest'
-import MarkdownView from '../../src/viewers/MarkdownView.js'
+import MarkdownView from '../../../src/components/viewers/MarkdownView.js'
 
 global.fetch = vi.fn()
